requirements.txt
alabaster==0.7.12
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asgiref==3.5.0
astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
binaryornot==0.4.4
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cookiecutter==1.7.3
cookiecutter-django==1.11.9
coverage==6.3.2
crispy-bootstrap5==0.6
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.12
django-allauth==0.49.0
django-coverage-plugin==2.0.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-model-utils==4.2.0
django-redis==5.2.0
django-rest-auth==0.9.5
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.13.1
docutils==0.17.1
executing==0.8.2
factory-boy==3.2.1
Faker==13.2.0
filelock==3.6.0
flake8==4.0.1
flake8-isort==4.1.1
hiredis==2.0.0
identify==2.4.11
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.0.1
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
lazy-object-proxy==1.7.1
livereload==2.6.3
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
poyo==0.5.0
pre-commit==2.17.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-slugify==6.1.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stack-data==0.2.0
termcolor==1.1.0
testfixtures==6.18.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
types-pytz==2021.3.5
types-PyYAML==6.0.4
typing-extensions==4.1.1
urllib3==1.26.8
virtualenv==20.13.2
watchdog==2.1.6
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
